Crate linera_alloy_genesis

Alloy genesis types

linera-alloy-genesis

Ethereum genesis file definitions.

The genesis file in an Ethereum blockchain is a JSON-formatted file used to define the initial state of the blockchain at the time of its creation.

Structs

ChainConfig

Defines core blockchain settings per block.

CliqueConfig

Consensus configuration for Clique.

EthashConfig

Empty consensus configuration for proof-of-work networks.

Genesis

The genesis block specification.

GenesisAccount

An account in the state of the genesis block.

ParliaConfig

Consensus configuration for Parlia. Parlia is the consensus engine for BNB Smart Chain. For the general introduction: https://docs.bnbchain.org/docs/learn/consensus/ For the specification: https://github.com/bnb-chain/bsc/blob/master/params/config.go#L558