Struct linera_alloy_genesis::ChainConfig

pub struct ChainConfig {

    pub chain_id: u64,
    pub homestead_block: Option<u64>,
    pub dao_fork_block: Option<u64>,
    pub dao_fork_support: bool,
    pub eip150_block: Option<u64>,
    pub eip150_hash: Option<B256>,
    pub eip155_block: Option<u64>,
    pub eip158_block: Option<u64>,
    pub byzantium_block: Option<u64>,
    pub constantinople_block: Option<u64>,
    pub petersburg_block: Option<u64>,
    pub istanbul_block: Option<u64>,
    pub muir_glacier_block: Option<u64>,
    pub berlin_block: Option<u64>,
    pub london_block: Option<u64>,
    pub arrow_glacier_block: Option<u64>,
    pub gray_glacier_block: Option<u64>,
    pub merge_netsplit_block: Option<u64>,
    pub shanghai_time: Option<u64>,
    pub cancun_time: Option<u64>,
    pub prague_time: Option<u64>,
    pub terminal_total_difficulty: Option<U256>,
    pub terminal_total_difficulty_passed: bool,
    pub ethash: Option<EthashConfig>,
    pub clique: Option<CliqueConfig>,
    pub parlia: Option<ParliaConfig>,
    pub extra_fields: BTreeMap<String, Value>,
    pub deposit_contract_address: Option<Address>,

}

Defines core blockchain settings per block.

Tailors unique settings for each network based on its genesis block.

Governs crucial blockchain behavior and adaptability.

Encapsulates parameters shaping network evolution and behavior.

See geth’s ChainConfig struct for the source of each field.

Fields
chain_id: u64

The network’s chain ID.

homestead_block: Option<u64>

The homestead switch block (None = no fork, 0 = already homestead).

dao_fork_block: Option<u64>

The DAO fork switch block (None = no fork).

dao_fork_support: bool

Whether or not the node supports the DAO hard-fork.

eip150_block: Option<u64>

The EIP-150 hard fork block (None = no fork).

eip150_hash: Option<B256>

The EIP-150 hard fork hash.

eip155_block: Option<u64>

The EIP-155 hard fork block.

eip158_block: Option<u64>

The EIP-158 hard fork block.

byzantium_block: Option<u64>

The Byzantium hard fork block (None = no fork, 0 = already on byzantium).

constantinople_block: Option<u64>

The Constantinople hard fork block (None = no fork, 0 = already on constantinople).

petersburg_block: Option<u64>

The Petersburg hard fork block (None = no fork, 0 = already on petersburg).

istanbul_block: Option<u64>

The Istanbul hard fork block (None = no fork, 0 = already on istanbul).

muir_glacier_block: Option<u64>

The Muir Glacier hard fork block (None = no fork, 0 = already on muir glacier).

berlin_block: Option<u64>

The Berlin hard fork block (None = no fork, 0 = already on berlin).

london_block: Option<u64>

The London hard fork block (None = no fork, 0 = already on london).

arrow_glacier_block: Option<u64>

The Arrow Glacier hard fork block (None = no fork, 0 = already on arrow glacier).

gray_glacier_block: Option<u64>

The Gray Glacier hard fork block (None = no fork, 0 = already on gray glacier).

merge_netsplit_block: Option<u64>

Virtual fork after the merge to use as a network splitter.

shanghai_time: Option<u64>

Shanghai switch time (None = no fork, 0 = already on shanghai).

cancun_time: Option<u64>

Cancun switch time (None = no fork, 0 = already on cancun).

prague_time: Option<u64>

Prague switch time (None = no fork, 0 = already on prague).

terminal_total_difficulty: Option<U256>

Total difficulty reached that triggers the merge consensus upgrade.

terminal_total_difficulty_passed: bool

A flag specifying that the network already passed the terminal total difficulty. Its purpose is to disable legacy sync without having seen the TTD locally.

ethash: Option<EthashConfig>

Ethash parameters.

clique: Option<CliqueConfig>

Clique parameters.

parlia: Option<ParliaConfig>

Parlia parameters.

extra_fields: BTreeMap<String, Value>

Additional fields specific to each chain.

deposit_contract_address: Option<Address>

The deposit contract address

Implementations

impl ChainConfig

pub fn is_homestead_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Homestead fork block.

pub fn is_eip150_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the EIP150 fork block.

pub fn is_eip155_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the EIP155 fork block.

pub fn is_eip158_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the EIP158 fork block.

pub fn is_byzantium_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Byzantium fork block.

pub fn is_constantinople_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Constantinople fork block.

pub fn is_muir_glacier_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Muir Glacier (EIP-2384) fork block.

pub fn is_petersburg_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Petersburg fork block.

pub fn is_istanbul_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Istanbul fork block.

pub fn is_berlin_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Berlin fork block.

pub fn is_london_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the London fork block.

pub fn is_arrow_glacier_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Arrow Glacier (EIP-4345) fork block.

pub fn is_gray_glacier_active_at_block(&self, block: u64) -> bool

Checks if the blockchain is active at or after the Gray Glacier (EIP-5133) fork block.

pub fn is_shanghai_active_at_block_and_timestamp( &self, block: u64, timestamp: u64, ) -> bool

Checks if the blockchain is active at or after the Shanghai fork block and the specified timestamp.

pub fn is_cancun_active_at_block_and_timestamp( &self, block: u64, timestamp: u64, ) -> bool

Checks if the blockchain is active at or after the Cancun fork block and the specified timestamp.